requirements.txt
--find-links https://download.pytorch.org/whl/cu118
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
scikit-learn==1.5.2
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.27.0
pillow==10.4.0
datasets==2.19.2
jupyterlab==4.2.5
torch==2.4.1
torchvision==0.19.1
torchsummary==1.5.1
flwr==1.11.0
flwr[simulation]
flwr-datasets==0.3.0
flwr_datasets[vision]
black==24.8.0
hydra-core==1.3.2
isort==5.13.2
pylint==3.3.1
pytest==8.3.3
docformatter==1.7.5
hydra-core==1.3.2
logging==0.4.9.6
imbalanced-learn==0.13.0